backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
mysql-connector-python
python-multipart
